requirements.txt
openai>=1.0.0
Pillow>=10.0.0
requests>=2.31.0
praw>=7.7.0
feedparser>=6.0.0
pytrends>=4.9.0
python-dotenv>=1.0.0
